refractor: seperated saveCode utility
